Fix small issues for CSV and JSON

+++ Runtime/JsonToFlexBufferConverter.cs
@@ -281,7 +281,7 @@ namespace FlexBuffers
 			}
 			else
 			{
-				flx.Add(int.Parse(builder.ToString()));
+				flx.Add(long.Parse(builder.ToString()));
 			}
 		}
